feat: Update to current version of the Amazon State Language (#292)
* Update StateMachine.j2119 language definition

This is current with version 0.7.0 of awslabs/statelint
https://github.com/awslabs/statelint

Further test cases to follow,

* Additional Fail state fixtures

* rejigged statelint_test

no need pass the explicit number of problems, because if there are any,
we know how many there are

* add jitter and delay tests

* add jitter and delay tests

* Added context-object tests

This unearthed a latent bug in @wmfs/j2119. PR raised
https://github.com/wmfs/j2119/pull/331
I've patch my local copy in the meantime so I can continue working

* Add map with parameters test

* dynamic timeout test

* Extend node probes to context object and choice states

Well, this is a big one. The map-with-null-itemspath showed
there were some changes to statelint itself that needed to be
ported across, specifically
* probeContextObjectAccess which checks some nullable and i
    non-nullable fields
* probeChoiceState which checks choice operators
and
* probePayloadBuilder, formerly probeParameters, which checks
    Parameters and ResultSelector

* Add resultselector tests

* Fix spelling in validation error message

* Function invocation tests

* Added task with static & dynamic heartbeat/timeout

* Further intrinsic function invocation tests

* Comment fields tests

* Further Map state tests

* Task credentials tests

* Update README to reference Sept 7 2023 version of ASL